chore: Move jobs to its own module
